+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project closely adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.2.5
+- Add - CONFIG now can save multiple snapshots in a stack
+- Add - features, scenarios and steps - add various metadata
+- Add - pytest-check for mulitple asserts
+- Add - steps - add browser info and screenshot info
+- Add - steps - add debug log and browser log
+- Add - steps - add section level, seq and parent_seq info
+- Change - extract generate_short_id to utils
+- Change - rename .start_timestamp => to .start_at
+- Change - use pathlib for paths
+- Chore - CI - fix results to not be deleted by pytest
+- Chore - makefile - make fix more like make lint
+
 ## 1.2.4
 - Fix - restore state after tab info
 - Change - replace offset with start time in step html report